one_d4: split RequestLivenessTest by concern
1,876 lines and 24 tests in one file meant one serial bazel target and
a whole-file blast radius whenever one timing-sensitive test flaked —
the submission-count race fixed in #1315 cost a full-suite rerun to
diagnose. Under java_test_suite each *Test.java is its own target, so
the file now splits on its three natural seams:

- RequestLeaseCeilingTest (9): the heartbeat and the run ceiling —
  slow runs keep renewing and vouching, wedged runs are cut loose,
  wedge-every-run requests are retired.
- RequestInterruptCheckpointTest (7): the month-boundary interrupt
  checkpoints and the interrupt handle's contract.
- RequestLeaseOwnershipTest (8): ownership under contention —
  renewals cannot revive the retired, taken leases fence writes, both
  entry points run a request once, mid-flight requests keep their
  dedupe slot.

RequestLivenessHarness (abstract, package-private, non-Test so the
suite compiles it into the shared lib) keeps the liveness story
javadoc, the constants, the clock/executor lifecycle, the claim and
worker builders, and every fake. Test bodies are unchanged — this is
a move, not a rewrite; the three targets now run in parallel and a
flake reruns only its own third.

All 8 worker targets green; the three new suites stable over
--runs_per_test=3.
