+++ domains/games/apis/one_d4/BUILD.bazel
@@ -402,7 +402,10 @@ java_test_suite(
         "src/test/java/com/muchq/games/one_d4/worker/IndexWorkerLifecycleTest.java",
         "src/test/java/com/muchq/games/one_d4/worker/IndexWorkerTest.java",
         "src/test/java/com/muchq/games/one_d4/worker/OpeningsTest.java",
-        "src/test/java/com/muchq/games/one_d4/worker/RequestLivenessTest.java",
+        "src/test/java/com/muchq/games/one_d4/worker/RequestInterruptCheckpointTest.java",
+        "src/test/java/com/muchq/games/one_d4/worker/RequestLeaseCeilingTest.java",
+        "src/test/java/com/muchq/games/one_d4/worker/RequestLeaseOwnershipTest.java",
+        "src/test/java/com/muchq/games/one_d4/worker/RequestLivenessHarness.java",
         "src/test/java/com/muchq/games/one_d4/worker/ResultMapperTest.java",
         "src/test/java/com/muchq/games/one_d4/worker/RetentionWorkerTest.java",
     ],
